Make error handling not just kill the process every time
I'm not sure what to do with errors. I'm pretty sure I don't want undelimited
continuations, and I certainly don't want to write C code in CPS. But dying on
every error is just inconvenient.

My current thinking is to emulate CL, have restarts. Fuck equational reasoning
I guess. So error returns a lispobj*, and an error handler could hypothetically
return a meaningful result on error.

Presently this causes problems. Some functions call error and ignore the
result, which is probably okay. One function, truth, well, it's not good. I
tried using it wrong and got a core dump. But it's some kind of start.

main.c, which is just the interpreter, now uses setjmp/longjmp in a handler to
work things. So now you can cause errors in the REPL and not kill everything.