Produce a runtime so instead of a flat executable
I renamed 'error' to 'lerror' because 'error' exists in libc and isn't marked
for TLS, so the linker balks even though our 'error' was different. A rename was
easier than trying to figure out shadowing.

main.c has been renamed repl.c and is written in terms of the runtime so.
Probably have to clean up the headers later.

Makefile is not ideal. Make is crazy. I kind of don't want to write a rule for
repl.o to avoid it being compiled with -fpic (which it is as of this commit,
pointlessly). Guess I shouldn't complain about make. What's professionalism?

The linker is also strange. -L. is probably not a good thing to have.