normalize object accessor definitions a bit
I wrote two more ambitious versions of this commit before this more incremental
one. The thing is that the C has more tag checks and general typelessness than
it needs to, but I'm not decided on how to go about removing it. Anyway, all
this macrology should make any future change a bit simpler.